Adicionando AdministradorDAO, e TesteCaseAdministrador

+++ biblioWeb/src/main/java/br/uepb/dao/usuarios/AdministradorDAO.java
@@ -10,18 +10,17 @@ import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 
 import br.uepb.dao.Conexao;
+import br.uepb.model.usuarios.Administrador;
 import br.uepb.model.usuarios.Funcionario;
-
 /**
- * Essa classe é responsável por se conectar com a tabela funcionario do Banco de Dados para operações de inserir, atualizar, remover e buscar objetos do tipo Funcionario. 
+ * Essa classe é responsável por se conectar com a tabela funcionario do Banco de Dados para operações de inserir, atualizar, remover e buscar objetos do tipo Administrador. 
  * @author EquipeACL
  *
  */
-public class FuncionarioDAO implements Interface_usuario<Funcionario> {
-
+public class AdministradorDAO {
 	private Connection con;
-	private static final Logger logger = LogManager.getLogger(FuncionarioDAO.class);
-	public FuncionarioDAO() {
+	private static final Logger logger = LogManager.getLogger(AdministradorDAO.class);
+	public AdministradorDAO() {
 		try {
 			con = Conexao.iniciarConexao();
 		} catch (Exception e) {
@@ -29,31 +28,31 @@ public class FuncionarioDAO implements Interface_usuario<Funcionario> {
 		}
 	}	
 	/**
-	 * Método para inserir um Funcionario no banco de dados
-	 * @param funcionario, objeto do tipo Funcionario
+	 * Método para inserir um Administrador no banco de dados
+	 * @param adm, objeto do tipo Administrador
 	 * @throws SQLException
 	 * @return true, se a inserção dos dados no banco de dados é bem sucedida
 	 * @return false, se houver algum problema durante a inserção dos dados no banco de dados
 	 */
-	public boolean createUsuario(Funcionario funcionario) {
+	public boolean createAdministrador(Administrador adm) {
 
 		PreparedStatement stmt = null;
 
 		try {
 			con = Conexao.iniciarConexao();
 			stmt = con.prepareStatement(
-					"insert into funcionario(cpf,rg,naturalidade,nomeCompleto,nomeUsuario,endereco,telefone,email,senha) values (?,?,?,?,?,?,?,?,?)");
+					"insert into funcionario(cpf,rg,naturalidade,nomeCompleto,nomeUsuario,endereco,telefone,email,senha,admin) values (?,?,?,?,?,?,?,?,?,?)");
 
-			stmt.setInt(1, funcionario.getCpf());
-			stmt.setInt(2, funcionario.getRg());
-			stmt.setString(3, funcionario.getNaturalidade());
-			stmt.setString(4, funcionario.getNomeCompleto());
-			stmt.setString(5, funcionario.getNomeUsuario());
-			stmt.setString(6, funcionario.getEndereco());
-			stmt.setInt(7, funcionario.getTelefone());
-			stmt.setString(8, funcionario.getEmail());
-			stmt.setString(9, funcionario.getSenhaAcesso());
-
+			stmt.setInt(1, adm.getCpf());
+			stmt.setInt(2, adm.getRg());
+			stmt.setString(3, adm.getNaturalidade());
+			stmt.setString(4, adm.getNomeCompleto());
+			stmt.setString(5, adm.getNomeUsuario());
+			stmt.setString(6, adm.getEndereco());
+			stmt.setInt(7, adm.getTelefone());
+			stmt.setString(8, adm.getEmail());
+			stmt.setString(9, adm.getSenhaAcesso());
+			stmt.setInt(10, 1);
 			stmt.executeUpdate();
 			return true;
 		} catch (SQLException e) {
@@ -73,19 +72,19 @@ public class FuncionarioDAO implements Interface_usuario<Funcionario> {
 	}
 
 	/**
-	 * Método para remover um Funcionario no banco de dados
-	 * @param funcionario, objeto do tipo Funcionario
+	 * Método para remover um Administrador no banco de dados
+	 * @param adm, objeto do tipo Administrador
 	 * @throws SQLException
 	 * @return true, se a inserção dos dados no banco de dados é bem sucedida
 	 * @return false, se houver algum problema durante a inserção dos dados no banco de dados
 	 */
-	public boolean removeUsuario(Funcionario funcionario) {
+	public boolean removeUsuario(Administrador adm) {
 		PreparedStatement stmt = null;
 
 		try {
 			con = Conexao.iniciarConexao();
-			stmt = con.prepareStatement("delete from funcionario where cpf=?");
-			stmt.setInt(1, funcionario.getCpf());
+			stmt = con.prepareStatement("delete from funcionario where cpf=? and admin = 1");
+			stmt.setInt(1, adm.getCpf());
 
 			stmt.executeUpdate();
 			return true;
@@ -105,28 +104,28 @@ public class FuncionarioDAO implements Interface_usuario<Funcionario> {
 	}
 
 	/**
-	 * Método para atualizar um Funcionario no banco de dados
-	 * @param funcionario, objeto do tipo Funcionario
+	 * Método para atualizar um Administrador no banco de dados
+	 * @param adm, objeto do tipo Administrador
 	 * @throws SQLException
 	 * @return true, se a inserção dos dados no banco de dados é bem sucedida
 	 * @return false, se houver algum problema durante a inserção dos dados no banco de dados
 	 */
-	public boolean updateUsuario(Funcionario funcionario) {
+	public boolean updateUsuario(Administrador adm) {
 		PreparedStatement stmt = null;
 
 		try {
 			con = Conexao.iniciarConexao();
 			stmt = con.prepareStatement("update funcionario set rg=?,naturalidade=?,nomeCompleto=?,nomeUsuario=?,endereco=?,telefone=?,email=?,senha=? where cpf=?");
 
-			stmt.setInt(1, funcionario.getRg());
-			stmt.setString(2, funcionario.getNaturalidade());
-			stmt.setString(3, funcionario.getNomeCompleto());
-			stmt.setString(4, funcionario.getNomeUsuario());
-			stmt.setString(5, funcionario.getEndereco());
-			stmt.setInt(6, funcionario.getTelefone());
-			stmt.setString(7, funcionario.getEmail());
-			stmt.setString(8, funcionario.getSenhaAcesso());
-			stmt.setInt(9, funcionario.getCpf());
+			stmt.setInt(1, adm.getRg());
+			stmt.setString(2, adm.getNaturalidade());
+			stmt.setString(3, adm.getNomeCompleto());
+			stmt.setString(4, adm.getNomeUsuario());
+			stmt.setString(5, adm.getEndereco());
+			stmt.setInt(6, adm.getTelefone());
+			stmt.setString(7, adm.getEmail());
+			stmt.setString(8, adm.getSenhaAcesso());
+			stmt.setInt(9, adm.getCpf());
 
 			stmt.executeUpdate();
 			return true;
@@ -145,27 +144,27 @@ public class FuncionarioDAO implements Interface_usuario<Funcionario> {
 	}
 
 	/**
-	 * Método para buscar um Funcionario no banco de dados
-	 * @param funcionario, objeto do tipo Funcionario
+	 * Método para buscar um Administrador no banco de dados
+	 * @param adm, objeto do tipo Administrador
 	 * @throws SQLException
-	 * @return ArrayList<Funcionario>, se houver um ou mais funcionarios que atendam ao requisito
+	 * @return ArrayList<Administrador>, se houver um ou mais administradores que atendam ao requisito
 	 * @return null, se a busca não retornar nenhum resultado	 
 	 */
-	public ArrayList<Funcionario> searchUsuario(Funcionario funcionario) {
-		ArrayList<Funcionario> listaFuncionario = new ArrayList<Funcionario>();
+	public ArrayList<Administrador> searchUsuario(Administrador adm) {
+		ArrayList<Administrador> listaAdministradores = new ArrayList<Administrador>();
 		PreparedStatement stmt = null;
 		ResultSet rs = null;
-		Funcionario f; 
+		Administrador  administrador; 
 		
 		try {
 			con = Conexao.iniciarConexao();
-			stmt = con.prepareStatement("select * from funcionario where nomeCompleto like ?");
-			stmt.setString(1,"%"+funcionario.getNomeCompleto()+"%");
+			stmt = con.prepareStatement("select * from funcionario where nomeCompleto like ? and admin = 1");
+			stmt.setString(1,"%"+adm.getNomeCompleto()+"%");
 			rs = stmt.executeQuery();
 			
 			while(rs.next()) {
-				f = new Funcionario(rs.getInt("cpf"), rs.getString("nomeCompleto"), rs.getInt("rg"), rs.getString("naturalidade"), rs.getString("endereco"), rs.getInt("telefone"), rs.getString("email"), rs.getString("senha"), rs.getString("nomeUsuario"));
-				listaFuncionario.add(f);
+				administrador = new Administrador(rs.getInt("cpf"), rs.getString("nomeCompleto"), rs.getInt("rg"), rs.getString("naturalidade"), rs.getString("endereco"), rs.getInt("telefone"), rs.getString("email"), rs.getString("senha"), rs.getString("nomeUsuario"));
+				listaAdministradores.add(administrador);
 			}
 			
 		} catch (SQLException e) {
@@ -181,7 +180,6 @@ public class FuncionarioDAO implements Interface_usuario<Funcionario> {
 			}
 		}
 		
-		return listaFuncionario;
+		return listaAdministradores;
 	}
-
 }
